+++ example/android/app/src/main/kotlin/com/transistorsoft/flutterbackgroundgeolocationexample/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.transistorsoft.flutterbackgroundgeolocationexample
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()