+++ ios/flutter_background_geolocation/Sources/flutter_background_geolocation/TSGeofencesChangeStreamHandler.m
@@ -1,0 +1,23 @@
+
+#import "./include/flutter_background_geolocation/TSGeofencesChangeStreamHandler.h"
+
+static NSString *const EVENT_NAME    = @"geofenceschange";
+
+@implementation TSGeofencesChangeStreamHandler
+
+- (NSString*) event {
+    return EVENT_NAME;
+}
+
+- (FlutterError*)onListenWithArguments:(id)arguments eventSink:(FlutterEventSink)events {
+    self.callback = ^void(TSGeofencesChangeEvent *event) {
+        events([event toDictionary]);
+    };
+    
+    [[TSLocationManager sharedInstance] onGeofencesChange:self.callback];
+    return nil;
+}
+
+@end
+
+