+++ ios/flutter_background_geolocation/Sources/flutter_background_geolocation/TSPowerSaveChangeStreamHandler.m
@@ -1,0 +1,24 @@
+
+#import "./include/flutter_background_geolocation/TSPowerSaveChangeStreamHandler.h"
+
+static NSString *const EVENT_NAME    = @"powersavechange";
+
+@implementation TSPowerSaveChangeStreamHandler
+
+- (NSString*) event {
+    return EVENT_NAME;
+}
+
+- (FlutterError*)onListenWithArguments:(id)arguments eventSink:(FlutterEventSink)events {
+    
+    self.callback = ^void(TSPowerSaveChangeEvent *event) {
+        events(@(event.isPowerSaveMode));
+    };
+    [[TSLocationManager sharedInstance] onPowerSaveChange: self.callback];
+    
+    return nil;
+}
+
+@end
+
+